:recycle: refactor: change service structure
