+++ src/main/java/com/grepp/coffee/app/service/MemberService.java
@@ -1,0 +1,5 @@
+package com.grepp.coffee.app.service;
+
+public class MemberService {
+
+}
