--- src/main/java/com/grepp/coffee/app/service/order/OrderServiceImpl.java
@@ -1,27 +1,0 @@
-package com.grepp.coffee.app.service.order;
-
-
-import com.grepp.coffee.app.model.dto.DetailedOrderDto;
-
-public class OrderServiceImpl implements OrderService {
-
-    @Override
-    public void processOrder(DetailedOrderDto detailedOrderDto) {
-
-    }
-
-    @Override
-    public void updateOrder(DetailedOrderDto detailedOrderDto) {
-
-    }
-
-    @Override
-    public void deleteOrder(DetailedOrderDto detailedOrderDto) {
-
-    }
-
-    @Override
-    public DetailedOrderDto getDetailedOrderDto() {
-        return null;
-    }
-}
